resolve component folder issue and refactor code
